ajustes tempo de inserção
