+++ Illumy.Oshi/pom.xml
@@ -5,6 +5,39 @@
     <artifactId>Illumy.Oshi</artifactId>
     <version>1.0-SNAPSHOT</version>
     <packaging>jar</packaging>
+    
+    <build>
+        <plugins>
+
+            <plugin>
+                <groupId>org.apache.maven.plugins</groupId>
+                <artifactId>maven-assembly-plugin</artifactId>
+                <version>2.4.1</version>
+                <configuration>
+                    <descriptorRefs>
+                        <descriptorRef>jar-with-dependencies</descriptorRef>
+                    </descriptorRefs>
+                    <archive>
+                      <manifest>
+                        <mainClass>pacote.subpacteSuaClasseComMetodoMain</mainClass>
+                      </manifest>
+                    </archive>
+
+                </configuration>
+                <executions>
+                  <execution>
+                    <id>make-assembly</id>
+                    <phase>package</phase>
+                    <goals>
+                        <goal>single</goal>
+                    </goals>
+                  </execution>
+                </executions>
+            </plugin>
+
+        </plugins>
+    </build>
+    
     <dependencies>
         <dependency>
             <groupId>com.microsoft.sqlserver</groupId>
