+++ Illumy.Oshi/src/main/java/com/mycompany/projetoillumy/oshi/InsertCPU.java
@@ -7,13 +7,11 @@ import org.springframework.jdbc.core.JdbcTemplate;
 
 public class InsertCPU {
     
-    protected Double usoCPU;   
-    protected String tempCPU;
+    protected Double usoCPU;       
     protected Integer fkServidor;
 
-    public InsertCPU(Double usoCPU,String tempCPU) {
-    this.usoCPU = usoCPU;                                  
-    this.tempCPU = tempCPU;                
+    public InsertCPU(Double usoCPU){
+    this.usoCPU = usoCPU;                                                     
     this.fkServidor = 1;                
     }
 
@@ -26,7 +24,7 @@ public class InsertCPU {
         try {
            jdbcTemplate.update(
                 "insert into CPU (utilizacaoAtual, TempCPU, fkServidor)"
-                + "values(?,?,?)", usoCPU, tempCPU,fkServidor);
+                + "values(?,?)", usoCPU,fkServidor);
              
         } catch (Exception e) {
             System.out.println("Erro"+e);
